Extract CTAButton from VotingCTABar
